refactor(node): inline intent resolution
